+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # Changelog
 
 ## Release 3.0.1
+
+### New Features
+
+### Breaking Changes
+
+* Support for Python 3.5, 3.6, 3.7, 3.8 are discontinued.
+* Switching to dotted-style options like in `pandas`. The old way of setting package-wide options
+  are removed. See documentation for more details.